path_shortener.py et al: identify & reproduce a bug!
I found a bug:

Due to how sorting is done, it appears, the namefile is being written to
disk before the folders are fully shortened, because some not-longest
folders with invalid chars require a namefile but were not sorted as the
longest, resulting in multiple namefiles for the same dirs.

A potential fix is to not write the namefiles until all renaming has
been done.

Alternatively, I could possibly re-sort the paths to fix while taking
into account illegal Winodows chars and namefiles that will need to be
created for trailing illegal dirs in long paths as a consequence.

I'm not sure yet what the fix will be.