config.py: decrease HASH_LEN, and fix path_shortener.py...
...to get the `path_shortener_demo.sh` test code to pass again!

I put the path_shortener.py code back to how it was, instead of trying
to incrementally decrease the HASH_LEN if the paths are still too long.

And, reducing the HASH_LEN from 4 to 3 was just enough to decrease the
path lengths for the `test_paths/` dir test with the new,
previously-increased PATH_LEN_ALREADY_USED config.py value.